CURL can throw errors when interacting with secure servers.
Add the --insecure option to the push_artifact Thor task.
Add error handling around the Thor task because we found a way for it to fail.
If --insecure is true, we will use the -k command with CURL to be insecure.
Switch to popen3 because stderr was surpressed.
Upping version number