Rework of configuration
The configuration of remote.rb was not behaving correctly. We don't need
to require a .nexus_cli if the user passes in overrides.
Updates to remote.rb and tasks.rb to properly set Remote.configuration in all cases.
Add the ability to override any Thor task.
Add color to all of get_nexus_status.
Upping to new minor version.